[EXE-2212] Adding JSON Functions PushDown Support in Snwoflake Connector
